refactor(file): scoped direct query解決を責務別に分割
